Refactor MapDrawer into partial classes
